django-app/requirements.txt
Django>=5.2,<5.3
django-allauth==65.9.0
djangorestframework==3.16.0
django-cors-headers==4.7.0
psycopg2-binary==2.9.10
gunicorn==23.0.0
redis==6.2.0
django-redis==5.4.0
django-rq==3.0.1
djangorestframework_simplejwt==5.5.0
python-dotenv==1.1.0
pillow==11.3.0
moviepy==2.2.1
whitenoise==6.9.0
pytest==8.4.1
pytest-django==4.11.1
coverage==7.10.0
